organize handlers by event type
