+++ patterns/fan-out-fan-in/task-service/infrastructure/stacks/lambda-stack.ts
@@ -162,7 +162,7 @@ export class LambdaStack extends cdk.Stack {
       functionName: `${props.appName}-list-tasks-${props.envName}`,
       runtime: lambda.Runtime.NODEJS_24_X,
       handler: 'handler',
-      entry: path.join(__dirname, '../../src/handlers/list-tasks.ts'),
+      entry: path.join(__dirname, '../../src/handlers/api/list-tasks.ts'),
       environment: lambdaEnvironment,
       timeout: cdk.Duration.seconds(10),
       memorySize: 256,
@@ -188,7 +188,7 @@ export class LambdaStack extends cdk.Stack {
       functionName: `${props.appName}-get-task-${props.envName}`,
       runtime: lambda.Runtime.NODEJS_24_X,
       handler: 'handler',
-      entry: path.join(__dirname, '../../src/handlers/get-task.ts'),
+      entry: path.join(__dirname, '../../src/handlers/api/get-task.ts'),
       environment: lambdaEnvironment,
       timeout: cdk.Duration.seconds(10),
       memorySize: 256,
@@ -214,7 +214,7 @@ export class LambdaStack extends cdk.Stack {
       functionName: `${props.appName}-create-task-${props.envName}`,
       runtime: lambda.Runtime.NODEJS_24_X,
       handler: 'handler',
-      entry: path.join(__dirname, '../../src/handlers/create-task.ts'),
+      entry: path.join(__dirname, '../../src/handlers/api/create-task.ts'),
       environment: lambdaEnvironment,
       timeout: cdk.Duration.seconds(10),
       memorySize: 256,
@@ -240,7 +240,7 @@ export class LambdaStack extends cdk.Stack {
       functionName: `${props.appName}-update-task-${props.envName}`,
       runtime: lambda.Runtime.NODEJS_24_X,
       handler: 'handler',
-      entry: path.join(__dirname, '../../src/handlers/update-task.ts'),
+      entry: path.join(__dirname, '../../src/handlers/api/update-task.ts'),
       environment: lambdaEnvironment,
       timeout: cdk.Duration.seconds(10),
       memorySize: 256,
@@ -266,7 +266,7 @@ export class LambdaStack extends cdk.Stack {
       functionName: `${props.appName}-delete-task-${props.envName}`,
       runtime: lambda.Runtime.NODEJS_24_X,
       handler: 'handler',
-      entry: path.join(__dirname, '../../src/handlers/delete-task.ts'),
+      entry: path.join(__dirname, '../../src/handlers/api/delete-task.ts'),
       environment: lambdaEnvironment,
       timeout: cdk.Duration.seconds(10),
       memorySize: 256,
@@ -292,7 +292,7 @@ export class LambdaStack extends cdk.Stack {
       functionName: `${props.appName}-create-task-subscriber-${props.envName}`,
       runtime: lambda.Runtime.NODEJS_24_X,
       handler: 'handler',
-      entry: path.join(__dirname, '../../src/handlers/create-task-subscriber.ts'),
+      entry: path.join(__dirname, '../../src/handlers/sqs/create-task-subscriber.ts'),
       environment: lambdaEnvironment,
       timeout: cdk.Duration.seconds(30),
       memorySize: 256,
@@ -333,7 +333,7 @@ export class LambdaStack extends cdk.Stack {
       functionName: `${props.appName}-upload-task-subscriber-${props.envName}`,
       runtime: lambda.Runtime.NODEJS_24_X,
       handler: 'handler',
-      entry: path.join(__dirname, '../../src/handlers/upload-task-subscriber.ts'),
+      entry: path.join(__dirname, '../../src/handlers/sqs/upload-task-subscriber.ts'),
       environment: lambdaEnvironment,
       timeout: cdk.Duration.seconds(60),
       memorySize: 512,
@@ -377,7 +377,7 @@ export class LambdaStack extends cdk.Stack {
       functionName: `${props.appName}-complete-taskfile-subscriber-${props.envName}`,
       runtime: lambda.Runtime.NODEJS_24_X,
       handler: 'handler',
-      entry: path.join(__dirname, '../../src/handlers/complete-task-file-subscriber.ts'),
+      entry: path.join(__dirname, '../../src/handlers/sqs/complete-task-file-subscriber.ts'),
       environment: lambdaEnvironment,
       timeout: cdk.Duration.seconds(30),
       memorySize: 256,
