+++ patterns/fan-out-fan-in/task-service/src/handlers/api/delete-task.test.ts
@@ -1,14 +1,12 @@
 import { APIGatewayProxyEvent, Context } from 'aws-lambda';
 
-import { Task } from '../models/task';
-
 // Mock dependencies BEFORE importing handler
-const mockGetTask = jest.fn();
+const mockDeleteTask = jest.fn();
 const mockLoggerInfo = jest.fn();
 const mockLoggerWarn = jest.fn();
 const mockLoggerError = jest.fn();
 
-jest.mock('../utils/config', () => ({
+jest.mock('../../utils/config', () => ({
   config: {
     TASKS_TABLE: 'test-tasks-table',
     AWS_REGION: 'us-east-1',
@@ -18,11 +16,11 @@ jest.mock('../utils/config', () => ({
   },
 }));
 
-jest.mock('../services/task-service', () => ({
-  getTask: mockGetTask,
+jest.mock('../../services/task-service', () => ({
+  deleteTask: mockDeleteTask,
 }));
 
-jest.mock('../utils/logger', () => ({
+jest.mock('../../utils/logger', () => ({
   logger: {
     info: mockLoggerInfo,
     warn: mockLoggerWarn,
@@ -32,14 +30,14 @@ jest.mock('../utils/logger', () => ({
   withRequestTracking: jest.fn(),
 }));
 
-describe('get-task handler', () => {
-  let handler: typeof import('./get-task').handler;
+describe('delete-task handler', () => {
+  let handler: typeof import('./delete-task').handler;
 
   beforeEach(() => {
     jest.clearAllMocks();
 
     // Import handler after mocks are set up
-    handler = require('./get-task').handler;
+    handler = require('./delete-task').handler;
   });
 
   const createMockEvent = (overrides?: Partial<APIGatewayProxyEvent>): APIGatewayProxyEvent => {
@@ -47,7 +45,7 @@ describe('get-task handler', () => {
       body: null,
       headers: {},
       multiValueHeaders: {},
-      httpMethod: 'GET',
+      httpMethod: 'DELETE',
       isBase64Encoded: false,
       path: '/tasks/123e4567-e89b-12d3-a456-426614174000',
       pathParameters: {
@@ -61,7 +59,7 @@ describe('get-task handler', () => {
         apiId: 'test-api-id',
         authorizer: null,
         protocol: 'HTTP/1.1',
-        httpMethod: 'GET',
+        httpMethod: 'DELETE',
         identity: {
           accessKey: null,
           accountId: null,
@@ -109,18 +107,9 @@ describe('get-task handler', () => {
   };
 
   describe('handler', () => {
-    it('should return task when it exists', async () => {
+    it('should return 204 when task is deleted successfully', async () => {
       // Arrange
-      const mockTask: Task = {
-        id: '123e4567-e89b-12d3-a456-426614174000',
-        title: 'Test Task',
-        detail: 'Test detail',
-        isComplete: false,
-        createdAt: '2025-11-01T10:00:00.000Z',
-        updatedAt: '2025-11-01T10:00:00.000Z',
-      };
-
-      mockGetTask.mockResolvedValue(mockTask);
+      mockDeleteTask.mockResolvedValue(true);
       const event = createMockEvent();
       const context = createMockContext();
 
@@ -128,15 +117,15 @@ describe('get-task handler', () => {
       const result = await handler(event, context);
 
       // Assert
-      expect(result.statusCode).toBe(200);
-      expect(JSON.parse(result.body)).toEqual(mockTask);
-      expect(mockGetTask).toHaveBeenCalledTimes(1);
-      expect(mockGetTask).toHaveBeenCalledWith('123e4567-e89b-12d3-a456-426614174000');
+      expect(result.statusCode).toBe(204);
+      expect(JSON.parse(result.body)).toEqual({});
+      expect(mockDeleteTask).toHaveBeenCalledTimes(1);
+      expect(mockDeleteTask).toHaveBeenCalledWith('123e4567-e89b-12d3-a456-426614174000');
     });
 
     it('should return 404 when task does not exist', async () => {
       // Arrange
-      mockGetTask.mockResolvedValue(null);
+      mockDeleteTask.mockResolvedValue(false);
       const event = createMockEvent();
       const context = createMockContext();
 
@@ -148,8 +137,8 @@ describe('get-task handler', () => {
       expect(JSON.parse(result.body)).toEqual({
         message: 'Task not found',
       });
-      expect(mockGetTask).toHaveBeenCalledTimes(1);
-      expect(mockGetTask).toHaveBeenCalledWith('123e4567-e89b-12d3-a456-426614174000');
+      expect(mockDeleteTask).toHaveBeenCalledTimes(1);
+      expect(mockDeleteTask).toHaveBeenCalledWith('123e4567-e89b-12d3-a456-426614174000');
     });
 
     it('should return 404 when taskId path parameter is missing', async () => {
@@ -167,7 +156,7 @@ describe('get-task handler', () => {
       expect(JSON.parse(result.body)).toEqual({
         message: 'Task not found',
       });
-      expect(mockGetTask).not.toHaveBeenCalled();
+      expect(mockDeleteTask).not.toHaveBeenCalled();
     });
 
     it('should return 404 when taskId is undefined', async () => {
@@ -185,13 +174,13 @@ describe('get-task handler', () => {
       expect(JSON.parse(result.body)).toEqual({
         message: 'Task not found',
       });
-      expect(mockGetTask).not.toHaveBeenCalled();
+      expect(mockDeleteTask).not.toHaveBeenCalled();
     });
 
     it('should return 500 error when service throws an error', async () => {
       // Arrange
       const mockError = new Error('Service error');
-      mockGetTask.mockRejectedValue(mockError);
+      mockDeleteTask.mockRejectedValue(mockError);
       const event = createMockEvent();
       const context = createMockContext();
 
@@ -201,22 +190,14 @@ describe('get-task handler', () => {
       // Assert
       expect(result.statusCode).toBe(500);
       expect(JSON.parse(result.body)).toEqual({
-        message: 'Failed to retrieve task',
+        message: 'Failed to delete task',
       });
-      expect(mockGetTask).toHaveBeenCalledTimes(1);
+      expect(mockDeleteTask).toHaveBeenCalledTimes(1);
     });
 
     it('should include CORS headers in response', async () => {
       // Arrange
-      const mockTask: Task = {
-        id: '123e4567-e89b-12d3-a456-426614174000',
-        title: 'Test Task',
-        isComplete: false,
-        createdAt: '2025-11-01T10:00:00.000Z',
-        updatedAt: '2025-11-01T10:00:00.000Z',
-      };
-
-      mockGetTask.mockResolvedValue(mockTask);
+      mockDeleteTask.mockResolvedValue(true);
       const event = createMockEvent();
       const context = createMockContext();
 
@@ -231,15 +212,7 @@ describe('get-task handler', () => {
 
     it('should log request context information', async () => {
       // Arrange
-      const mockTask: Task = {
-        id: '123e4567-e89b-12d3-a456-426614174000',
-        title: 'Test Task',
-        isComplete: false,
-        createdAt: '2025-11-01T10:00:00.000Z',
-        updatedAt: '2025-11-01T10:00:00.000Z',
-      };
-
-      mockGetTask.mockResolvedValue(mockTask);
+      mockDeleteTask.mockResolvedValue(true);
       const event = createMockEvent();
       const context = createMockContext();
 
@@ -250,43 +223,10 @@ describe('get-task handler', () => {
       // Handler execution is verified through successful test completion
     });
 
-    it('should return task with only required fields', async () => {
-      // Arrange
-      const mockTask: Task = {
-        id: '123e4567-e89b-12d3-a456-426614174000',
-        title: 'Test Task',
-        isComplete: false,
-        createdAt: '2025-11-01T10:00:00.000Z',
-        updatedAt: '2025-11-01T10:00:00.000Z',
-      };
-
-      mockGetTask.mockResolvedValue(mockTask);
-      const event = createMockEvent();
-      const context = createMockContext();
-
-      // Act
-      const result = await handler(event, context);
-
-      // Assert
-      expect(result.statusCode).toBe(200);
-      const body = JSON.parse(result.body);
-      expect(body).toEqual(mockTask);
-      expect(body).not.toHaveProperty('detail');
-      expect(body).not.toHaveProperty('dueAt');
-    });
-
     it('should handle different task IDs', async () => {
       // Arrange
-      const taskId = 'different-task-id';
-      const mockTask: Task = {
-        id: taskId,
-        title: 'Different Task',
-        isComplete: true,
-        createdAt: '2025-11-01T10:00:00.000Z',
-        updatedAt: '2025-11-01T10:00:00.000Z',
-      };
-
-      mockGetTask.mockResolvedValue(mockTask);
+      const taskId = 'different-task-id-789';
+      mockDeleteTask.mockResolvedValue(true);
       const event = createMockEvent({
         pathParameters: { taskId },
         path: `/tasks/${taskId}`,
@@ -297,9 +237,48 @@ describe('get-task handler', () => {
       const result = await handler(event, context);
 
       // Assert
-      expect(result.statusCode).toBe(200);
-      expect(mockGetTask).toHaveBeenCalledWith(taskId);
-      expect(JSON.parse(result.body)).toEqual(mockTask);
+      expect(result.statusCode).toBe(204);
+      expect(mockDeleteTask).toHaveBeenCalledWith(taskId);
+    });
+
+    it('should return 204 with empty body on successful deletion', async () => {
+      // Arrange
+      mockDeleteTask.mockResolvedValue(true);
+      const event = createMockEvent();
+      const context = createMockContext();
+
+      // Act
+      const result = await handler(event, context);
+
+      // Assert
+      expect(result.statusCode).toBe(204);
+      expect(result.body).toBe('{}');
+    });
+
+    it('should log the taskId when deletion succeeds', async () => {
+      // Arrange
+      mockDeleteTask.mockResolvedValue(true);
+      const event = createMockEvent();
+      const context = createMockContext();
+
+      // Act
+      await handler(event, context);
+
+      // Assert
+      // Handler execution is verified through successful test completion
+    });
+
+    it('should log the taskId when task not found', async () => {
+      // Arrange
+      mockDeleteTask.mockResolvedValue(false);
+      const event = createMockEvent();
+      const context = createMockContext();
+
+      // Act
+      await handler(event, context);
+
+      // Assert
+      // Handler execution is verified through successful test completion
     });
   });
 });
