+++ patterns/fan-out-fan-in/task-service/src/handlers/api/get-task.ts
@@ -1,8 +1,8 @@
 import { APIGatewayProxyEvent, APIGatewayProxyResult, Context } from 'aws-lambda';
 
-import { getTask } from '../services/task-service.js';
-import { internalServerError, notFound, ok } from '../utils/apigateway-response.js';
-import { logger, withRequestTracking } from '../utils/logger.js';
+import { getTask } from '../../services/task-service.js';
+import { internalServerError, notFound, ok } from '../../utils/apigateway-response.js';
+import { logger, withRequestTracking } from '../../utils/logger.js';
 
 /**
  * Lambda handler for retrieving a task by ID
