+++ patterns/fan-out-fan-in/task-service/src/handlers/api/list-tasks.test.ts
@@ -1,13 +1,13 @@
 import { APIGatewayProxyEvent, Context } from 'aws-lambda';
 
-import { Task } from '../models/task';
+import { Task } from '../../models/task';
 
 // Mock dependencies BEFORE importing handler
 const mockListTasks = jest.fn();
 const mockLoggerInfo = jest.fn();
 const mockLoggerError = jest.fn();
 
-jest.mock('../utils/config', () => ({
+jest.mock('../../utils/config', () => ({
   config: {
     TASKS_TABLE: 'test-tasks-table',
     AWS_REGION: 'us-east-1',
@@ -17,11 +17,11 @@ jest.mock('../utils/config', () => ({
   },
 }));
 
-jest.mock('../services/task-service', () => ({
+jest.mock('../../services/task-service', () => ({
   listTasks: mockListTasks,
 }));
 
-jest.mock('../utils/logger', () => ({
+jest.mock('../../utils/logger', () => ({
   logger: {
     info: mockLoggerInfo,
     error: mockLoggerError,
